piece together chunked messages
websockets don't guarantee that the entire original message will be sent
in one go.  this patch uses the length in the QAP header to determine if
the entire message has been received.

as long as the message is not complete, it holds the frames in an array.
once the sum of the sizes of the frames equals the length in the QAP
header plus the header size, it reassembles the messages and passes it on.