allow named oob messages in ocap mode
instead of switching on opts.ocap_mode, switch on the type of the wrapped
payload's first element:

1. if it's a string, check if on_oob_message supplied. if so, call it
with the usual function(error, result) callback.

NOTE: the signature was the opposite before, function(message, error),
so this is a breaking change!  but (error, result) is the standard
asynchronous javascript signature, so i think this is correct.

2. if it's a function, make sure we are in ocap mode. if so, call the function
exactly as before.

3. if it's anything else, report "Unknown oob message type".